Update-Some Quality of life features are here
Addition: Template feature, allows a string to be inputted as a template(examples provided/usable in comboBox), then outputted based from the given count(Default is 1). It inserts at last mouse cursor location in the editingTextarea. This is for the editingWindow.
Addition: CTRL-Z and CTRL-Y, aka Undo and Redo are now functional on the editing window. (Cut/Copy-Paste works with hotkeys with no changes, no rightclick though)

Change: All widgets but the editingTextArea are now in their own frame for better organization.